requirements.txt
dm-haiku==0.0.6
jax==0.3.4
jaxlib==0.3.2+cuda11.cudnn82
optax==0.1.1
numpy==1.22.3
Pillow==9.0.1
scipy==1.8.0
setuptools==45.2.0
typing_extensions==4.1.1
gym==0.23.1
gym-notices==0.0.6
dm-control==1.0.0
tensorflow==2.8.0
tensorflow-datasets==4.5.2
tf-agents==0.12.0